Fix issue where get_dbx_c_stack_trace can do a ps on 0 and print an error message into test output
The jnl_crash/C9B11-001794 test failed in internal testing because a process was stuck with a blocking pid of 0. This caused com/gtmprocstuck_get_stack_trace.csh to call com/get_dbx_c_stack_trace.csh which passed the blocking pid to a ps process. The ps process returned an `error: process ID out of range` message which was printed into the test's output log causing a test failure. As it is possible for the dbx script to be invoked with a blocking pid of 0, this commit bypasses the ps process in that case and instead sets pid_alive to 0.

This commit also removes references to an obsolete architecture from the jnl_crash test's instream and outref files. The jnl_crash/C9B11-001794 test was disabled on this architecture which is no longer supported by YottaDB, the upstream software or even by the company that created the architecture.