Fix for online_integ/ointeg4 in case online integ detects kill-in-progress
The issue here is a small window in ss_initiate.c between the wait loop that no KILLs are in progress
and when the process re-obtains crit, it is possible for a KILL to sneak in, grab crit and have its
phase one commit complete then the online integ re-obtains crit blocking the second phase KILL commit
until it completes resulting in MUKILLIP warnings and related block busy type messages. These messages
must be filtered out from the online integ output before the framework sees them and fails the test.

Note this fix takes the logic to do this filtering and puts it in a new script
com/filter_mukillip_from_oli_output.csh