[V63009] New v63009/gtm9037 test to make sure that an error message appears if journaling has to be stopped due to an error
This commit adds a v63009/gtm9037 test that ensures that errors that can shut down journaling or cause an instance freeze result in a %YDB-E-JNLEXTEND message in the syslog. The test sets $ydb_error_on_jnl_file_lost and then causes an error with permissions when autoswitching journal files. This produces a %YDB-E-JNLEXTEND error in the syslog which does not show up without the gtm9037 changes in V6.3-009.