[V63009] Address Review Comments
This commit addresses various review comments relating to new tests added as part of the V6.3-009 merge.

The r132/ydb652 test has 2 changes. Firstly, an inefficient routine has been removed and replaced by setting a local variable with subscripts to replace calls to the routine and checking that local variable to get the needed values. Secondly, the verify routine is changed so that the test will fail if an ibase other than 8 or 16 gets passed to it. The same change has been made to the v63005/gtm5574 test from which this test was derived.

The v63009/gtm9155 test has been modified to move the various M test routines into a gtm9155.m file rather than generating 7 different M files at runtime.

The r132/ydb441 test now mentions in the output that the %YDB-E-ZLINKFILE error is expected.

There are also readability changes to the test output for v63009/gtm9037 and other tests.