Fix rare failure in manually_start/gds_max_blk due to an edge case.
The manually_start/gds_max_blk test failed on rare occasions in internal testing. The cause of this failure is due to $pid2extend, $pid2integ and $pid2size not being set due to $i never being equal to $rand_under in a rare circumstance.

In the test, there is a loop that runs from $i=1 until $i=$max (495). In this loop, it sets $pid2extend, $pid2integ and $pid2size when $i=$rand_under. $rand_under is set equal to $random($thres) and $thres always equals ($max*.88)+1 (436.6). Since $random returns an integer that can be 0, it was possible for $rand_under to be randomly chosen as 0. If this happened, $i would never equal $rand_under and after the loop finished the test would check the value of $pid2integ which was not set.

The fix removes the possibility for $rand_under to be set at 0 while keeping the maximum value of $rand_under the same (435 since the .6 is truncated when $random converts 436.6 to an integer). This guarantees that $pid2extend, $pid2integ and $pid2size will be set when the test expects them to be.