Fix r124/ydb312_gtm8182a test failure involving extra lines of output
The r124/ydb312_gtm8182a test failed in internal testing because of an issue with an extra line of output containing a mysterious line containing a 5 printed after the correct output to `number of processes attached to JNLPOOL of INST3: ` but before the blank line printed by jnlprocnum.csh. We believe the failure to have been caused by the grep command run on the output from an `ipcs -a` command matching an extra line due to a fluke. It was impossible to confirm this however as the `ipcs -a` command's output is not saved anywhere so there was insufficient information to be certain of this.

This consists of 3 fixes:

1) The command is now `ipcs -m` instead of `ipcs -a` to reduce the amount of extra information included in the output. The test only needs the line of output relating to a specific shared memory id so removing the extra information should not cause any problems.
2) The output of the `ipcs -m` command is now recorded to make it easier to analyze failures if this test fails again in the future.
3) The test now matches the shared memory ids against only the second column while ensuring that the first column contains `m` as expected followed by a single space. This is intended to prevent incorrect lines being matched because the shmid happens to be contained in another column of the output.