Fix manually_start_1/4g_journal and v63009_1/gtm9037replic test failures
This commit fixes 2 test failures from internal testing last weekend.

The manually_start_1/4g_journal test failed on an ARMv7L machine due to exceeding the time limit of 5 seconds for `MUPIP JOURNAL -show=header`. It took 17 seconds on one run for the `show=header` to finish due to hardware slowness. The fix for this failure is to increase the limit from 5 to 60 seconds.

The v63009_1/gtm9037replic test failed on an x86_64 machine running Ubuntu 18.04 because the test ran too quickly. The test is supposed to cause an instance freeze in the receiver and crash its update process by removing write access on its journal directory. It then restores write access, restarts the update process and unfreezes the receiver in order to shut down replication. However, it can occasionally restore write access and attempt to bring back the update process and unfreeze the receiver too quickly if the receiver has not filled up its journal and frozen itself yet. The fix for this is to wait for a %YDB-E-REPLINSTFROZEN message in the syslog before attempting to bring back the frozen receiver to give the receiver enough time to fill up its journal file and freeze itself because it can't write a new journal file.