[YDB#635] MUPIP INTEG (no options) ignores Ctrl-C
Adds an expect test that issues a CTRL-C to the mupip integ code. Checks
that mupip integ doesn't produce any additional output other than File
or Region in the expect output file once Ctrl-C is triggered. This
verifies that Ctrl-C terminated the program.

Added an additional condition in v63004/gtm8909 test to ensure that
mupip help which exits with exit status of 1 doesn't fail. The current
implementation of mupip returns with exit(1) status when CTRL+C is
pressed. The condition accounts for this change.