Fix v63009_1/gtm9037replic failure due to instance freeze in other tests
At one point, the v63009_1/gtm9037replic test waits for an YDB-E-REPLINSTFROZEN message to show up in the syslog before attempting to unfreeze the instance and restart the receiver's update process. This failed in internal testing due to another test running at the same time doing an instance freeze which got written to the syslog causing the v63009_1/gtm9037replic test to attempt to unfreeze the instance and restart the receiver's update process before the instance freeze happened. The fix is to wait for an "Update process exiting" message in the update process log instead.