Fix various <sudo> subtest reference files since ydbinstall.sh now overwrites pkg-config file if YDB version is same
Previously ydbinstall.sh used to skip the step where `pkg-config` file is updated if the to-be-installed
YDB version is the same as the version in the `pkg-config` file.

But this was modified as part of YDB@cfb3f43c613dd89edb04989fa001af9b0e63ed7c to update even if the YDB
version is the same.

This causes a different output to show up in various `sudo` subtest reference files (all of which use
the `ydbinstall.sh` script) if the YottaDB version used to run the test is newer than (or identical to)
the one previously recorded in `pkg-config`.

All these reference files files are now updated to allow the new type of output along with the previous
type of output (which is still possible if the YottaDB version used to run the test is older than the
one recorded in `pkg-config`).