Fix v63009/gtm9037replic failures due to RENAMEFAIL errors showing up in update process log file
The v63009/gtm9037replic test failed in internal testing due to a RENAMEFAIL error message that showed up in the syslog. This can happen depending if the test removes write permissions on the receiver process's journal directory at the same time that the receiver is renaming a journal file. If this happens, a RENAMEFAIL will end up in the update process's log file and the test will fail. This is more likely to happen on pro builds because of how the set operations are optimized and is also more likely to happen on faster machines because journal files will fill up more quickly. This commit fixes the failure by filtering RENAMEFAILs out of the update process log file.