[Ubuntu 20.04] Bypass/exclude tests with a V5 version dependency where those versions had to be removed
Ubuntu 20.04 has an incompatibility with GT.M versions earlier than V62000. Because of this the
version V61000 and all V5 versions have been removed from servers that were upgraded to 20.04.
It appears though that Ubuntu 20.10 does not have this incomaptibility so some servers that were
upgraded directly from 18.04 to 20.10 still have their older versions and run just fine.

These changes are to identify those servers that were upgraded to 20.04 and had the versions
removed and set a test system flag so subtests with V5 dependencies are excluded.