[YDB#663] New r132/ydb663 subtest to verify DBDUPNULCOL error is not incorrectly issued
This commit adds a new r132/ydb663 test that ensures that loading a binary extract back into the same database with at least 1 null subscript does not produce a YDB-E-DBDUPNULCOL error. Prior to the YDB#663 code changes, this would produce such an error for both standard null collation and historical null collation due to a bug discovered while creating the v63009/gtm9144 test.

This commit also fixes the v63009/gtm9144 test to print the output of the MUPIP LOADs as it happens. Previously, both were redirected to the same file which meant that the test output contained only 1 YDB-E-DBDUPNULCOL. Between them, r132/ydb663 and v63009/gtm9144 provide test coverage for loading binary extracts into a database with all 4 combinations of null subscript collation types ensuring that the YDB-E-DBDUPNULCOL error shows up when it is supposed to and does not show up when it is not supposed to.