[YDB#612] Assertion failure in -direct mode with suspend/resume
Test uses expect to send SIGSTOP to YottaDB and then foregrounds it
again using fg on the shell, which is what caused the issue to manifest.

Two additional tests were suggested by @nars1 and added to illustrate
the effects of the issue on pro builds.
1. echo is turned off again when we go back to YottaDB
2. Canonical mode is turned on again when we go back to YottaDB.

Tests have been updated to prevent fg from causing race conditions by
adding sleeps in the expect script.