Fix v63009/gtm9134 test failures on Red Hat and CentOS involving extra line of output
The v63009/gtm9134 test failed on Red Hat and CentOS machines due to an extra line of output that was sent to stderr by the fake source server (/dev/urandom) that sends spurious messages to the receiver. All output sent by this process to stdout was discarded by redirecting it to /dev/null but output to stderr was not redirected. This produced test failures because Red Hat and CentOS machines produced an extra `Ncat: Connection reset by peer.` message in stderr which ended up in the test output. As the output of the command may be useful for debugging future test failures, output from that process to both stderr and stdout is now redirected to a file called nc_stderr.out. With stderr redirected, the test now passes on CentOS and Red Hat machines.