[YDB#673] Changes to test output due to YDB#673, add new test r132/ydb673
The vast bulk of these changes are due to two things:
1. The 32->64 bit hash change which changed the sizes of control blocks thus affecting
LOCKSPACEINFO messages.
2. The increase of default lock space from 40 to 220 pages also changed the LOCKSPACEINFO
messages as well as DSE file header output and LOCKSPACEUSE messages.

Also in this update are a number of changes to r124/ydb297 (YDB#673 reverts ydb297 but
we retain the testing) so it can pass with the new code. There are similar changes in
r126/ydb438. But r126/ydb439 had to be disabled as it uses $ydb_lockhash_n_bits to restrict
the number of bits in a hash value which leads us to LOCKSPACEFULL messages that we currently
cannot avoid (we increased the hash size precisely for this reason).

Note the LOCKSPACEINFO message changes have two values - one for the 32 bit builds and one for
64 bit builds.

Note the v63000/gtm8511 test had a numeric conversion removed (+ operator in front of extract
expression) because the extract expression was actually a hex value. Previous usages in this
test correctly extracted the correct value with no hex digits. With these changes, the offset
value had an initial 0 then a hex digit so the value became 0 instead of the offset it was
supposed to be. Removing the '+' from the equation, stored the value as a string instead
allowing the correct value to be preserved.

Also in this update is the new r132/ydb673 test that tests that we can still get all the
locks even though the lock list will hash to the same list so a rehash/resize is required.
This tests new code added for ydb673 to retry a lock again if it failed due to needing
garbage collection, rehash, or resize even if the timeout was 0.