[YDB#676] New r132/ydb676 test to verify that mupip journal -rollback -fetchresync handles spurious input correctly
This commit adds a new r132/ydb676 test along the lines of the v63009/gtm9134 test that sends spurious input from /dev/urandom to a `mupip journal -rollback -fetchresync` command to verify that the command ignores the spurious input and functions correctly. The test revealed that the spurious message issue that affected replication receiver servers is also an issue for `mupip journal -rollback -fetchresync` commands.