[YDB#685][#341] New v63009/gtm9115 test to verify correctness and V6.3-009 performance improvement for conversion utilities
This commit adds a new v63009/gtm9115 test based heavily on the r130/ydb485 test that tests $HO, %OH, %DO and %OD for correctness and improved performance compared to the pre V6.3-009 versions in a similar manner to how r130/ydb485 tests %HD and %DH. The test also tests the YDB#685 changes. Portions of the test that test YDB#685 are marked accordingly in the reference file. This commit also makes some changes to the r130/ydb485 test to speed it up by reducing various 15 second stages to 2 seconds.