[YDB#682] New r132/ydb682 test to make sure %HO converts hexadecimal numbers to the same octal number whether or not the input has a "0x" prefix.
This commit adds a r132/ydb682 test that converts randomly generated hexadecimal numbers to octal for 2 seconds. Each randomly generated number is converted to octal without a prefix, with a `Ox` prefix and with a `0X` prefix and the outputs are compared to make sure that the `0x` or `0X` prefix is ignored by the %HO utility.