[YDB#685][#341] Use CPU time instead of elapsed time for timing/gtm9115 and r130/ydb485 tests
This commit addresses test failures in the performance comparison section of the timing/gtm9115 and r130/ydb485 tests by switching them to use 2 seconds of CPU time instead of 2 seconds of elapsed time. It also treats anything above 90% of the previous implementation as a pass for the 14 digit %HO comparison because the new implementation, which is usually slightly above 100% of the previous implementation's performance, came in at 94.72% once in internal testing out of 100 runs using CPU time.