Update threeenp1C2.m to print message when forked off worker processes terminate normally
Not all worker processes farmed out by the threeenp1C2 main show evidence of ending either
by signal or by panic in the go/fatal_signal test. The question is whether some of these
workers finished normally or not. Since the workers do not print a termination message when
they finish, we do not know. This adds a simple termination message to the output.

In addition changed an os.Exit(0) call to a simple return so the defer chain is driven as
the process shuts down. Previously, this was not happening causing potential database
issues like one failure seen in go/fatal_error test where we failed in rundown with the
following assert failure:

%YDB-F-ASSERT, Assert failed in /Distrib/YottaDB/V998_R131/sr_unix/wcs_flu.c line 703 for expression (FALSE && "not in known white box test")