[YDB#697] New r132/ydb697 test to verify correctness of CONVNEG^%CONVBASEUTIL for octal input
This test verifies the correctness of CONVNEG^%CONVBASEUTIL for octal input. It is heavily based on the v63005/gtm5574 test and operates by generating random octal numbers from 1 to 10 digits and verifying that these numbers are converted to the right two's complement number by CONVNEG. It also pads randomly generated numbers with 0's up to a total of 10 digits to ensure that these numbers are being converted correctly as well. The original v63005/gtm5574 failure was extremely difficult to reproduce as it only affected a handful of random 5 digit negative numbers that were incorrectly converted from decimal to octal with %DO due to this bug in CONVNEG.