[#336] V63009 Add test for GTM-8921 - Tests that IPCs for read-only databases are correctly run down when last process exits
The crux of the changes for this issue were that the IPCs were correctly cleaned up where
they previously were not. This is verified with this test which fails with V63008 based
versions and passes with V63009 based versions. Ran it 30 times just to make sure random
testsystem selections didn't mess with it. Also verified it fails 100% on V63008.