[V63010] Fix relink/zlink test failure due to GTM-9185 changes
This commit fixes the relink/zlink test which was failing around 40-50% of the time as described below by Steve Estes:

```What this does is tell the validation code not to expect a cycle change. This is due to GTM-9185 which changes how the relink works when a path is specified for the input source. GT.M (and now YDB) no longer obey $ZROUTINES rules for object placement when the source is specified instead of found via $ZROUTINES rules. This piece of this test specifically compiles a version of a routine and expected it to be loaded in that same directory according to ZLINK rules but that no longer happens so we eliminate the cycle change it was expecting.```