[V63010] Fix outref for manually_start/gds_max_blk test
The manually_start/gds_max_blk test failed because the outref wasn't updated to reflect V6.3-010 changes to `DSE DUMP -BLOCK` output. This commit fixes that. This outref update had been missed previously because the manually_start test is part of the M suite which isn't run in E_ALLs.