Fix v63010/gtm9206 failure where $gtm_chset is undefined
This commit fixes a test failure in v63010/gtm9206 where the test failed is $gtm_chset was randomly set to undefined rather than explicitly set to M or UTF-8. This caused the check for UTF-8 mode to fail with an undefined variable error. The fix is to check if $gtm_chset is defined before checking if it is set to UTF-8.