[#378] New v63010/gtm9180 test for error when DSE block numbers are larger than largest 32 big signed int
This commit adds a v63010/gtm9180 test that checks that `DSE -add` and `DSE -dump` produce error messages as expected when the block number is larger than the largest 32 bit signed integer. Upstream version V6.3-010 contains changes to DSE that prepared for a future expansion of block sizes to 64 bit in upstream version V7.0-000. The error message should go away and cause this test to fail once V7.0-000 is integrated. The test only runs under a pro build because larger block numbers result in an assert failure and core on dbg builds.

There is also a fix to the instream for both the v63009 and v63010 tests. These contained duplicate blocks to exclude tests from running under pro. One of these blocks has been removed.