Fix r130/ydb485 weekend test failure on ARMV7L by disabling performance checks on ARMVXL
The r130/ydb485 test failed on an ARMV7L machine during the weekend tests. The part that failed was the performance check which tests that the current implementations of %DH and %HD are faster than the pre-r1.30 versions. This had consistently passed until the tests were shortened from 15 second intervals to 2 second intervals. The intervals were increased back to 5 seconds after the previous such failure back in February. As the purpose of these tests is to ensure that the performance is faster, we would expect much more frequent failures if there was a real issue and the tests have always passed (so far) on x86 machines with 5 second intervals, we decided to disable the performance checks on ARMVXL instead of increasing the intervals and dramatically increasing the runtime of this test.