[#378] Fix issues with v63010/gtm9180 test when gtmdbglvl is randomly set
One of the environment variables that is randomly set by the test system's com/do_random_settings.csh is $gtmdbglvl which enables certain debugging settings. When it is set to 0x40000, freed storage is automatically set to 0xdeadbeef to identify use after free memory bugs. In sr_unix/mucblkini.c, the BSTAR_REC() macro on line 96 does a SET_CMPC() macro to set the value to 0 when it creates a pointer. When it is called with a value of 0, it only sets the 3rd byte of the header, leaving the 4th as garbage.

As this is only an issue for the tests, we decided to modify the outref to ignore the differences in output on a handful of `dse -dump` commands. This commit implements this change to the outref and also removes the line from the u_inref that ensured that gtmdbglvl was unset.

While working on this fix for v63010/gtm9180, we discovered a number of other tests where a `##TEST_AWK` contained `|` characters that were not escaped or within `()`. This commit also fixes these and updates those outref files to reflect their output.