[#377] New v63010/gtm9178 test checks for ERRWZTIMEOUT on $ZTIMEOUT that executes a runtime error on timeout
In upstream versions V6.3-006 to V6.3-009, running a `$ztimeout` command from a direct mode prompt in a terminal where the ztimeout vector (code executed on timeout) resulted in a runtime error would cause the upstream software to crash with a `GTMASSERT2` failure. In V6.3-010, this assert failure was replaced with an `ERRWZTIMEOUT` error message. Due to YDB#712, this assert failure did not appear in any YottaDB versions. Instead, YottaDB returned an appropriate error message from the runtime error. This test runs an expect script to verify that the `ERRWZTIMEOUT` message appears as expected.