[YDB#712] Fix intermittent r132/ydb712 test failures dealing with shell prompts
The initialization code sets the SHELL variable to the current shell so should be used anytime
the expected output value is the shell prompt. This test did not originally do this so even though
the shell prompt was always recognized by expect, sometimes the shell prompt did not end up in the
expect output file. But by expecting "SHELL", instead of ">", we still get a match and the expected
SHELL output always ends up in the reference file.