Disable r130/ydb485 performance checks section on aarch64
The r130/ydb485 test failed this past weekend on an aarch64 machine. Previously, we disabled the performance checks on ARMV6L and ARMV7L machines due to similar failures and had increased the intervals from 2 to 5 to 8 seconds to address failures on x86 machines. Rather than increase the intervals again and further increase the runtime of the test, this commit disables the r130/ydb485 performance checks on aarch64 machines.