[YDB#612] Attempt to fix rare expect timing failures
More expect shenanigans.

Previously, we had these lines in our expect script
(r132/u_inref/ydb612.exp):

```
1 # Prompt makes it easy to just figure out that we are on a shell
2 spawn /usr/local/bin/tcsh -f
3 send -- "set prompt='# '\r"
4 # Expect regex checks that we are on a new line with a '# '
5 expect -re {(?n)^# }
```

In the output, we previously had this:
```
3  spawn /usr/local/bin/tcsh -f
4  set prompt='# '
5  > set prompt='# '
```

On one of the machines running Arch Linux, line 4 disappeared during one
of the tests, and we have never been able to repeat it since then.

My only explanation is that the shell was starting (3) and was
(perhaps) in a mode which did not permit echoing, which is why line 4
didn't print. We don't check that the shell has actually started, since
we really don't know what text to check for (that's the whole point of
setting "prompt").

Using `com/expectinit.exp` will supposedly fix this problem, although I
am not entirely convinced, as I think the same condition could happen
there. In any case, I do think it's less likely that we see this over
there as there are a lot more reads/writes.