New r132/ydb737 subtest to test that Call-in APIs know how to handle $QUIT
Test for MR YDB!974.

The C program's output previously looked like this before the fix:

```
Failure of quit1-1 with error: 150374554,(Call-In),%YDB-E-QUITARGREQD, Quit from an extrinsic must have an argument
Failure of quit1-2 with error: 150374554,(Call-In),%YDB-E-QUITARGREQD, Quit from an extrinsic must have an argument
```

This is incorrect. The correct output after the fix is:

```
quit1 return is supposed to be 5: 5
quit1 return is supposed to be 150373210: 150373210
```