[YDB#739] New r132/ydb739 test ensures that MUPIP INTEG -SUBSCRIPT does not SIG-11 with NULL subscripts and no end key specified.
This commit adds a r132/ydb739 test that runs a MUPIP INTEG -SUBSCRIPT on a database with null subscripts where the end key is not specified. This MUPIP INTEG command produced a SIG-11 on YottaDB versions r1.30 and earlier. According to the V6.3-010 release note tested for by #371, this should have been fixed in V6.3-010 but we discovered that it wasn't and that our test still produces a SIG-11 as of upstream version V6.3-014.