[#371] New v63010/gtm8322 test to test MUPIP SIZE -SUBSCRIPT
This commit adds a new v63010/gtm8322 test that tests the MUPIP SIZE -SUBSCRIPT qualifier. This qualifier was added in upstream version V6.3-010 and specifies the range of globals, which can be subscripted, for MUPIP SIZE to estimate. It is incompatible with MUPIP SIZE -SELECT and this test verifies that using both -SELECT and -SUBSCRIPT produces an error. The test also runs the MUPIP SIZE -SUBSCRIPT equivalent of a command that caused MUPIP INTEG -SUBSCRIPT to SIG-11 without the YDB#739 changes which produces an error message.