[YDB#742] Enhance call_ins/test_ci_z_halt_rc to test zhalt 0, zhalt return values > 255, zhalt return values < 0
- zhalt 0 is supposed to behave like halt.
- YDB code previously didn't allow for zhalt returns greater than 255.

Additional changes:
- Don't print zstatus if nothing to print. This keeps the lines of
specific tests together.
- zhalt [code] produces an error status, however, $zstatus is not
populated. However, it may contain a value from a previous failing
scenario, creating the situation where we think that zhalt created that
failure; but it actually didn't, the zstatus was there from a previous
failure. Now, we reset $zstatus at the beginning of each test to ensure
that this doesn't happen. Ref files updated accordingly.