[YDB#704] New r132/ydb704 test that invoking YottaDB via valgrind works
This commit adds a r132/ydb704 test that runs several valgrind commands to ensure that a valgrind process that invokes YottaDB does not produce a YDBDISTUNVERIF error. This happened due to valgrind changing the process name from `YOTTADB` to `MEMCHECK-` followed by the platform.

This commit also changes the imptp-rs pipeline to use the latest-master branch instead of the latest branch for the debian docker image. This fixes an imptp-rs pipeline failure caused by the latest branch using an outdated version of cargo that caused a compile error when building YDBRust.