[YDB#744] Update mpt_extra/testdate test to reflect changes to CDS^%H
After the recent changes to CDS^%H to display dates after 2/28/2100 rather than an empty string, the mpt_extra/testdate test failed because it was expecting the presence of this limitation. This commit updates the mpt_extra/testdate test to expect CDS^%H to return the correct date rather than an empty string for dates after 2/28/2100.