[#368] New v63010/gtm1044 test for new $ZATRANSFORM optional 3rd arguments
This commit adds a v63010/gtm1044 test which tests the new optional $ZATRANSFORM 3rd arguments `2` and `-2`. These arguments take the first character of a string in M mode and return either the next or the previous collated character. This commit also adds the required gtm_ac_xutil() functions to all existing collations in the test system. Without these, the new $ZATRANSFORM arguments will cause $ZATRANSFORM to return a YDB-E-COLLATIONUNDEF. This also tests that the expected YDB-E-ZATRANSCOL is returned when these new $ZATRANSFORM arguments are used in UTF-8 mode.