[YDB#704] Disable r132/ydb704 test on ARMVXL
This commit disables the r132/ydb704 test on ARMVXL because the test requires valgrind which is not available for 32-bit ARM platforms. It also adjusts the setup dependencies to check for the presence of valgrind on x86_64 and AARCH64 systems. This is to ensure that we don't see any r132/ydb704 failures due to valgrind not being installed.