Fix sudo/plugins failures by renaming $utf8 variable to $utf
The sudo/plugins test failed occassionally with the wget error `Error during wget of YottaDB distribution file https://gitlab.com/api/v4/projects/7957109/repository/tags/`. The cause of the failure was due to the test using a variable name ($utf8) that is sometimes set in `com/set_gtmroutines.csh`.

When `com/set_gtmroutines.csh` set $utf8 to `/utf8`, this overrode the $utf8 variable in the test which was set to `--utf8 default` which confused the ydbinstall script into thinking that the test should install a version of YottaDB called `utf8`. When it attempted to wget this version from gitlab, the wget command would fail because no `utf8` tag exists on gitlab in the YDB repo. The fix for this is to rename the variable in the sudo/plugins test from $utf8 to $utf to eliminate the conflict with the test system's $utf8 variable.