Fix sudo/ydb306 failures and update outref for sudo/plugins and sudo/pluginsonly
This commit fixes a couple failures for the sudo/ydb306 test:

* The test was failing because the reference file was never updated after the YDB!1012 change that added "Now installing YDBZlib" to the output for the `--zlib` option.
* On an in-house Arch machine, the test would sometimes fail with a `NONUTF8LOCALE` error because the locale was automatically set to `ANSI_X3.4-1968`. The fix is to do the `$switch_chset "UTF-8"` earlier which ensures that a UTF-8 locale is set. This is the same thing that is done in the sudo/plugins and sudo/pluginsonly tests to avoid `NONUTF8LOCALE` errors.

This commit also updates the sudo/plugins and sudo/pluginsonly tests to reflect the change in YDB!1020 that suppresses the `Cloning into '.'...` message. This change was made because this message did not show up on an in-house RHEL7 machine causing these tests to fail.