Fix spurious timing/gtm9115 failures by tolerating up to 10% slowdown for 20 digit %DO
The timing/gtm9115 test failed because one run of 20 digit %DO was only 99.6% as fast as the previous implementation. This was the first time it had failed. This commit modifies the test to treat up to a 10% slowdown for 20 digit %DO as a pass just as !1239 and !1149 did for other parts of this test.