Address v60003/gtm7777 failures on slow ARM machines by raising flush_time from 5 minutes to 1 hour
The v60003/gtm7777 test has been failing pretty regularly on slow ARM machines of late due to what appears to be extra, unexpected flushes in case 2. This causes the DFL, DFS and/or JFS fields to be different than expected. The test currently sets the `flush_time` to 5 minutes. This commit increases it to the maximum of 1 hour which should hopefully fix these failures.