Fix v54003/C9L06003421 test failures on ARM by allowing up to 10 seconds time gap (not 5); Reduce time gap from 5 to 2 seconds on x86_64
The v54003/C9L06003421 test has been failing on ARM machines in the weekend test with a `Mismatch in timestamp` error in the part when comparing the EOF timestamp to the tp_resolve time towards the end of the test. In all failure instances, the time gap has been 7 seconds. This commit fixes these failures by allowing a time gap of up to 10 seconds on ARM machines.

As we discovered that the 5 second time gap is excessive for x86_64, it is reduced from 5 to 2 seconds. We had tested reducing it to 1 second but that produced 2 failures out of 500. At 2 seconds, we got 0 failures out of 500.

This commit also fixes an error with the `copyright.py` script that caused the end of the `All rights reserved.` line to be misaligned when automatically updating copyrights. The line had used spaces instead of tabs.