Fix v52000/C9H03002835 test to check if a socket OPEN command timed out
The v52000/C9H03002835 test failed in internal testing after an attempt to open a socket resulted in a timeout. This revealed that the test did not check if the attempt to open the socket succeeded and just outputted "Connection to receiver established" whether or not the connection was established.

As an `OPEN` command that times out before successfully establishing a connection with a device sets `$test` to `0` (meaning `FALSE`), the fix is to check if `$test` is `0` immediately after the `OPEN` command. If it is (meaning the `OPEN` timed out), the test fails and immediately halts with an appropriate error message. Otherwise (meaning the `OPEN` succeeded), the test writes the `Connection to receiver established` message and continues.