Address v62000/gtm7824 failure by moving heavy computations before the mupip integ
The v62000/gtm7824 test failed on an ARMV6L machine that was under heavy load due to running another test at the same time. In the test, there is a `MUPIP INTEG` command that waits for 30 seconds due to a white box test so that the INTEG process can set a flag allowing the `waitforOLIstart.csh` process to proceed. In the failure, the `offset.csh` script that is run after the integ starts but before the `waitforOLIstart.csh` took 39 seconds which meant that the integ finished before `waitforOLIstart.csh` started, causing the test to fail. The fix is to run `offset.csh` before the integ so that the integ will not terminate early on slower machines under heavy load.