[YDB#793] Enhance r130/ydb566 to also test for ydb793 fix to ydb566 bug
This commit enhances the existing r130/ydb566 test to also test 2 additional scenarios for external call tables:

* Comments at the beginning of the external call table before the shared library line
* Comments at the end of the shared library line

Both of these cases did not work in external call tables prior to ydb793 and are fixed by ydb!1043. As the call in table does not use the shared library line and as the part of the test that tests comments in call in tables already has a comment on its first line, there is no need to change the call in table part of the test.