[YDB#803] Enhance expect/D9F05002548 to test for HOME/END keys
The tests duplicate the tests that contain CTRL-A/CTRL-E and makes sure
that they still work. A few were not duplicated, as ESC gets special
handling as a default terminator, and thus there were instances where
HOME/END key couldn't exactly replace CTRL-A/CTRL-E. An example is that
in READ NOEDITING mode, CTRL-A does not terminate a read, whereas any
ESC sequence (including HOME/END) does.